test: Move `resolveTestVariantsConfig` to global lib
