+++ node/packages/aws-lambda-sdk/test/integration/index.test.js
@@ -14,7 +14,7 @@ const { default: fetch } = require('node-fetch');
 const cleanup = require('../lib/cleanup');
 const createCoreResources = require('../lib/create-core-resources');
 const processFunction = require('../lib/process-function');
-const resolveTestVariantsConfig = require('../lib/resolve-test-variants-config');
+const resolveTestVariantsConfig = require('../../../../test/lib/resolve-test-variants-config');
 const resolveFileZipBuffer = require('../utils/resolve-file-zip-buffer');
 const awsRequest = require('../../../../test/utils/aws-request');
 const pkgJson = require('../../package');
