+++ node/test/lib/resolve-test-variants-config.js
